Add a Learn Mode; save the learn state to a volume
In Learn Mode, MMM shows a symbol an the user has to type it in morse.
The symbols are chosen from a "learning set": initially containing only
the vowels, but new symbols are added as the user masters other symbols.

Among the symbols in the learning set, there's a higher probability of
symbols not yet very well mastered to be selected.